pyproject.toml
[tool.poetry]
name = "buffett"
version = "0.1.0"
description = "AI stock trader"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
openai = "^1.42.0"
tushare = "^1.4.8"
workalendar = "^17.0.0"
python-dotenv = "^1.0.1"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
ratelimit = "^2.2.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
